backend/services/data_service/requirements.txt
fastapi==0.110.2
uvicorn[standard]==0.29.0

python-dotenv==1.0.1
python-jose[cryptography]==3.3.0
httpx[http2]==0.27.0
python-multipart==0.0.9

sqlalchemy==2.0.30
pymysql==1.1.1

boto3==1.34.99
PyMuPDF==1.23.21
celery==5.3.1
redis==4.5.5